Projeto 1
- Unificação dos dataobject access para possibilitar de forma coerente o
conceito de transação
- CreateAtleta executa insertAtletaHistorico, a transação é unica tanto
para os dados do atleta como do seu historico
- troca de NUMBER(3,5) para FLOAT no script
- DENY_RESULT_INSERT ainda não testado
- scripts de população de dados considerando "cod".nextval

+++ sql_scripts/script_Create_Tables_And_COD_SEQUENCES.sql
@@ -55,7 +55,7 @@ CREATE SEQUENCE codAtleta MINVALUE 1 INCREMENT BY 1 START WITH 1;
 CREATE TABLE SerieClassificacao(
     codSerie NUMBER NOT NULL,
     codAtleta NUMBER NOT NULL,
-    marca NUMBER(3,5),
+    marca FLOAT,
     raiaParticipacao NUMBER,
     colocacao NUMBER,
     statusClassificacaoAtleta NVARCHAR2(15) 
@@ -69,9 +69,9 @@ CREATE TABLE AtletaHistorico(
     codAtleta NUMBER NOT NULL,
     codTorneio NUMBER NOT NULL,
     codModalidade NUMBER NOT NULL,
-    marca NUMBER(3,5) NOT NULL,
+    marca FLOAT NOT NULL,
     colocacao NUMBER NOT NULL,
-    nota NUMBER(3,5) NULL
+    nota FLOAT NULL
 );
 ALTER TABLE AtletaHistorico ADD CONSTRAINT FKATLETAHISTORICO_ATLETA 
     FOREIGN KEY (codAtleta) REFERENCES Atleta (codAtleta);
